보스 이동 높낮이 조절

+++ Source/UnrealPortfolio/GAS/GATA/GATA_BossStruggleSquare.h
@@ -18,10 +18,10 @@ public :
 	AGATA_BossStruggleSquare();
 
 	UPROPERTY(EditAnywhere)
-	int32 BoxSizeXValue;
+	float BoxSizeXValue;
 
 	UPROPERTY(EditAnywhere)
-	int32 BoxSizeYValue;
+	float BoxSizeYValue;
 
 protected :
 	virtual void GetAttributeSetting() override;
